app/requirements.txt
mysql-connector-python==8.0.33
tabulate==0.8.10
python-dotenv==0.21.0
